Polynomial template needs separate domain_t, range_t.
I fixed a conceptual mistake in the Polynomial class template.
It really needs separate types for domain and range, just
like the Interpolator class template.

For example, the caller might need a polynomial that
maps a real-valued `x` to a complex-valued `y`.
The previous version of this code forced `x` to be complex
if `y` was complex, forcing the inefficiency of converting
a real value to complex when evaluating poly(x).