Bug fix: was not handling multiplication of empty Polynomials.
When the coefficient list inside a Polynomial object is empty,
the Polynomial should be treated as representing f(x) = 0.
There was a crash trying to multiply two such empty polynomials,
because I was trying to create a vector containing -1 elements.
This threw a std::range_error.
Reproduced the problem with a unit test, then fixed the bug.