Support polynomial composition. Allow 0^0 = 1.
Implemented the ability to compose polynomials: h(x) = f(g(x)).

Instead of throwing an exception for Polynomial 0 raised to
the integer power 0, treat 0^0 = 1. This is consistent
with the concept of composing the function f(x) = 1 = x^0
with g(x) = 0:

    f(g(x)) = (0)^0 = 1

Any caller that wants to disallow 0^0 can do so via

    if (poly.isZero() && exponent == 0)
        ...